// src/static/enums/app.modals.ts
enum APP_MODALS {
	EXAMPLE_MODAL = '3X4MPL3_M0D4L',
	CONTACT_MODAL = 'CONTACT_MODAL',

	LOGIN_MODAL = 'LOGIN_MODAL',

	IMAGE_MODAL = 'IMAGE_MODAL',
	DELETE_MODAL = 'DELETE_MODAL',
	DELETE_PROJECT = 'DELETE_PROJECT',
	DELETE_COMMENT = 'DELETE_COMMENT',

	SESSION_EXPIRED = 'SESSION_EXPIRED',
}

export default APP_MODALS;